Complete factory pattern basic tests
